+++ cerebro/sources/lco.py
@@ -9,19 +9,19 @@
 from __future__ import annotations
 
 import asyncio
+import re
 import warnings
-from contextlib import suppress
-from datetime import datetime, timedelta
-from sqlite3 import OperationalError
+from datetime import datetime
 
-from typing import Any, Awaitable, Callable, Optional
+from typing import Any, ClassVar, Optional
 
-import peewee
+import httpx
+from polars import date
 
-from cerebro.sources.source import DataPoints
+from sdsstools.utils import cancel_task
 
-from .. import log
-from .source import Source
+from cerebro import log
+from cerebro.sources.source import Source
 
 
 warnings.filterwarnings(
@@ -31,10 +31,8 @@ warnings.filterwarnings(
 )
 
 
-class LCOWeather(Source):
-    """Connects to the LCO weather database and grabs some information.
-
-    Connection parameters are hardcoded here because they are unlikely to change.
+class DIMMSource(Source):
+    """Retrieve DIMM data from the LCO ``clima.lco.cl`` API.
 
     Parameters
     ----------
@@ -42,6 +40,8 @@ class LCOWeather(Source):
         The name of the data source.
     bucket
         The bucket to write to. If not set it will use the default bucket.
+    route
+        The route to the API endpoint.
     tags
         A dictionary of tags to be associated with all measurements.
     interval
@@ -49,147 +49,88 @@ class LCOWeather(Source):
 
     """
 
-    source_type = "lco_weather"
+    source_type: ClassVar[str] = "lco_dimm_data"
+    base_url: ClassVar[str] = "http://clima.lco.cl"
+
     interval: float = 90
 
     def __init__(
         self,
         name: str,
         bucket: Optional[str] = None,
+        route: str = "dimm_state",
         tags: dict[str, Any] = {},
         interval: float | None = None,
     ):
         super().__init__(name, bucket, tags)
 
         self.interval = interval or self.interval
+        self.route = route
 
-        self.connection = peewee.MySQLDatabase(
-            "weather",
-            user="lcodata",
-            host="clima.lco.cl",
-            read_default_file="~/.my.cnf",
-        )
-
-        self.last_data: datetime = datetime.utcnow() - timedelta(minutes=10)
-
-        self._runner: asyncio.Task | None = None
-        self._tasks: list[Callable[[], Awaitable]] = [
-            self.dimm_data,
-            self.magellan_data,
-        ]
+        self._runner_task: asyncio.Task | None = None
+        self._last_data = datetime.now(date.timezone.utc)
 
     async def start(self):
         """Starts the runner."""
 
-        if self._runner:
+        if self._runner_task:
             await self.stop()
 
-        self._runner = asyncio.create_task(self._run_tasks())
+        self._runner_task = asyncio.create_task(self._get_dimm_data())
 
         await super().start()
 
     async def stop(self):
         """Stops the runner."""
 
-        if self._runner:
-            with suppress(asyncio.CancelledError):
-                self._runner.cancel()
-                await self._runner
-
+        self._runner_task = await cancel_task(self._runner_task)
         self.running = False
 
-    async def _run_tasks(self):
-        """Runs the defined tasks."""
+    async def _get_dimm_data(self):
+        """Gets DIMM data from the API."""
 
         while True:
-            data = []
+            try:
+                async with httpx.AsyncClient(base_url=self.base_url) as client:
+                    response = await client.get(self.route)
+                    response.raise_for_status()
+                    data = response.text
 
-            if self.connection.is_closed():
-                conn_result = False
-                try:
-                    conn_result = self.connection.connect()
-                except OperationalError as err:
-                    log.error(f"{self.name}: failed connecting to database: {err!s}")
-
-                if not conn_result:
-                    log.error(f"{self.name}: failed connecting to database.")
-
-            if not self.connection.is_closed():
-                results = await asyncio.gather(
-                    *[asyncio.wait_for(task(), 20) for task in self._tasks],
-                    return_exceptions=True,
+                match = re.match(
+                    r"^time = (?P<time>.+)\naz = (?P<az>.+)\nel = (?P<el>.+)\n"
+                    r"seeing = (?P<seeing>.+)\ncounts = (?P<counts>.+)\n$",
+                    data,
                 )
 
-                for ii, result in enumerate(results):
-                    coro_name = self._tasks[ii].__name__
-                    if isinstance(result, asyncio.CancelledError):
-                        log.error(f"{self.name}: task {coro_name} timed out.")
-                    elif isinstance(result, Exception):
-                        log.error(f"{self.name}: task {coro_name} failed: {result!s}.")
-                    else:
-                        data += result
+                if not match:
+                    raise ValueError("Could not parse DIMM data API response.")
 
-            if len(data) > 0:
-                data_points = DataPoints(data=data, bucket=self.bucket)
-                self.on_next(data_points)
+                match_dict = match.groupdict()
+                time = datetime.fromisoformat(match_dict["time"] + "Z")
+                alt = float(match_dict["el"])
+                seeing = float(match_dict["seeing"])
 
-            self.connection.close()
-            self.last_data = datetime.utcnow()
+                # Avoid adding the same point again and again during the day.
+                if time <= self._last_data:
+                    await asyncio.sleep(self.interval)
+                    continue
 
-            await asyncio.sleep(self.interval)
+                self._last_data = time
 
-    async def dimm_data(self):
-        """Gathers data from DIMM."""
+            except Exception as ee:
+                log.error(f"Failed to get DIMM data: {ee}")
+                await asyncio.sleep(self.interval)
+                continue
 
-        if self.connection.is_closed():
-            raise RuntimeError("Database connection is not open.")
-
-        time_str = self.last_data.strftime("'%Y-%m-%d %H:%M:%S'")
-        qdata = self.connection.execute_sql(
-            f"SELECT tm, se, el FROM dimm_data WHERE tm > {time_str} ORDER BY tm DESC"
-        )
-
-        points = [
-            {
-                "measurement": "dimm",
-                "fields": {"seeing": vv[1], "altitude": vv[2]},
-                "time": vv[0],
-                "tags": self.tags.copy(),
-            }
-            for vv in qdata
-        ]
-
-        return points
-
-    async def magellan_data(self):
-        """Gathers data from the Magellans."""
-
-        if self.connection.is_closed():
-            raise RuntimeError("Database connection is not open.")
-
-        time_str = self.last_data.strftime("'%Y-%m-%d %H:%M:%S'")
-        qdata = self.connection.execute_sql(
-            "SELECT tm, un, fw, az, el FROM magellan_data "
-            f"WHERE fw IS NOT NULL AND fw > 0 AND tm > {time_str}"
-            "ORDER BY tm DESC"
-        )
-
-        points = []
-        tags = self.tags.copy()
-
-        for vv in qdata:
-            if vv[1] == 0:
-                telescope = "baade"
-            else:
-                telescope = "clay"
-
-            points.append(
-                {
-                    "measurement": "magellan",
-                    "fields": {"seeing": vv[2], "azimuth": vv[3], "altitude": vv[4]},
-                    "time": vv[0],
-                    "tags": {**tags, "telescope": telescope},
-                }
+            self.on_next(
+                [
+                    {
+                        "measurement": "dimm",
+                        "fields": {"seeing": seeing, "altitude": alt},
+                        "time": time,
+                        "tags": self.tags.copy(),
+                    }
+                ]
             )
 
-        return points
+            await asyncio.sleep(self.interval)